src/python/v2/requirements.txt
aiohttp==3.8.5
PyYAML==6.0.1